refactor folder structure for EV
