+++ src/ensembl/src/content/app/entity-viewer/EntityViewer.tsx
@@ -18,12 +18,12 @@ import {
 import { toggleSidebar } from 'src/content/app/entity-viewer/state/sidebar/entityViewerSidebarActions';
 
 import { StandardAppLayout } from 'src/shared/components/layout';
-import GeneView from './components/views/gene-view/GeneView';
-import EntityViewerAppBar from 'src/content/app/entity-viewer/components/entity-viewer-app-bar/EntityViewerAppBar';
-import EntityViewerSidebar from './components/entity-viewer-sidebar/EntityViewerSideBar';
-import EntityViewerSidebarTabs from './components/entity-viewer-sidebar-tabs/EntityViewerSidebarTabs';
-import EntityViewerSidebarToolstrip from './components/entity-viewer-sidebar-toolstrip/EntityViewerSidebarToolstrip';
-import EntityViewerTopbar from './components/entity-viewer-topbar/EntityViewerTopbar';
+import GeneView from './gene-view/GeneView';
+import EntityViewerAppBar from './shared/components/entity-viewer-app-bar/EntityViewerAppBar';
+import EntityViewerSidebar from './gene-view/components/entity-viewer-sidebar/EntityViewerSideBar';
+import EntityViewerSidebarTabs from './gene-view/components/entity-viewer-sidebar-tabs/EntityViewerSidebarTabs';
+import EntityViewerSidebarToolstrip from './shared/components/entity-viewer-sidebar-toolstrip/EntityViewerSidebarToolstrip';
+import EntityViewerTopbar from './shared/components/entity-viewer-topbar/EntityViewerTopbar';
 
 import { RootState } from 'src/store';
 import { EnsObject } from 'src/shared/state/ens-object/ensObjectTypes';
