+++ src/ensembl/src/content/app/entity-viewer/gene-view/GeneView.tsx
