+++ src/ensembl/src/content/app/entity-viewer/gene-view/components/base-pairs-ruler/basePairsRulerHelper.ts
