+++ src/ensembl/src/content/app/entity-viewer/gene-view/components/entity-viewer-sidebar/EntityViewerSideBar.tsx
