+++ src/ensembl/src/content/app/entity-viewer/shared/components/entity-viewer-app-bar/EntityViewerAppBar.tsx
