+++ src/ensembl/src/content/app/entity-viewer/shared/components/entity-viewer-sidebar-toolstrip/EntityViewerSidebarToolstrip.tsx
