+++ src/ensembl/stories/entity-viewer/base-pairs-ruler/BasePairsRuler.stories.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useRef } from 'react';
 import { storiesOf } from '@storybook/react';
 
-import BasePairsRuler from 'src/content/app/entity-viewer/components/base-pairs-ruler/BasePairsRuler';
+import BasePairsRuler from 'src/content/app/entity-viewer/gene-view/components/base-pairs-ruler/BasePairsRuler';
 
 import styles from './BasePairsRuler.stories.scss';
